Reorganization and minor fixes.
Move SingleExtensionFits and MultiExtensionFits to processors.
Move primary header stuff to FitsProcessor as common pattern.
Fix minor problems with priority values.
Move FitsProcessor to UploadProcessor level.
