+++ trail/upload/process_uploads/fits_processor.py
@@ -12,18 +12,12 @@ import matplotlib.pyplot as plt
 
 import astropy.visualization as aviz
 from astropy.io import fits
-from astropy.io.fits import ImageHDU
-from astropy.io.fits.hdu.image import PrimaryHDU
-from astropy.io.fits.hdu.compressed import CompImageHDU
 
-from django.db import transaction
-
-from upload.models import Frame
 from upload.process_uploads.upload_processor import UploadProcessor
 from upload.process_uploads.header_standardizer import HeaderStandardizer
 
 
-__all__ = ["FitsProcessor", "SingleExtensionFits", "MultiExtensionFits"]
+__all__ = ["FitsProcessor",]
 
 
 class FitsProcessor(UploadProcessor):
@@ -44,6 +38,9 @@ class FitsProcessor(UploadProcessor):
     def __init__(self, upload):
         super().__init__(upload)
         self.hdulist = fits.open(self._upload.tmpfile.temporary_file_path())
+        self.primary = self.hdulist["PRIMARY"].header
+        self.standardizer = HeaderStandardizer.fromHeader(self.primary,
+                                                          filename=upload.filename)
         self.isMultiExt = len(self.hdulist) > 1
 
     @staticmethod
@@ -212,132 +209,3 @@ class FitsProcessor(UploadProcessor):
         and inserts it into the database.
         """
         raise NotImplemented()
-
-
-class SingleExtensionFits(FitsProcessor):
-
-    name = "SingleExtensionFits"
-    priority = 1
-
-    def __init__(self, upload):
-        super().__init__(upload)
-        primary = self.hdulist["PRIMARY"].header
-        self.standardizer = HeaderStandardizer.fromHeader(primary, filename=upload.filename)
-        self.imageData = self.hdulist["PRIMARY"].data
-
-    @classmethod
-    def canProcess(cls, upload):
-        canProcess, hdulist = super().canProcess(upload, returnHdulist=True)
-        return canProcess and not cls._isMultiExtFits(hdulist)
-
-    def standardizeWcs(self):
-        return self.standardizer.standardizeWcs()
-
-    def standardizeHeaderMetadata(self):
-        return self.standardizer.standardizeMetadata()
-
-    def standardizeHeader(self, short=True):
-        standardizedWcs = {"wcs" : self.standardizeWcs()}
-        standardizedMetadata = {"metadata" : self.standardizeHeaderMetadata()}
-        if short:
-            standardizedMetadata.update(standardizedWcs)
-        else:
-            standardizedMetadata = dict(standardizedMetadata["metadata"],
-                                        **standardizedWcs["wcs"])
-        return standardizedMetadata
-
-    @transaction.atomic
-    def storeHeaders(self, short=True):
-        frame = Frame(**self.standardizeHeader(short=False))
-        frame.save()
-
-    def storeThumbnails(self):
-        large, small = self._createThumbnails(self._upload.basename, self.imageData)
-        plt.imsave(small["savepath"], small["thumb"], pil_kwargs={"quality": 10})
-        plt.imsave(large["savepath"], large["thumb"], pil_kwargs={"quality": 30})
-
-
-class MultiExtensionFits(FitsProcessor):
-
-    name = "MultiExtensionFits"
-    priority = 1
-
-    def __init__(self, upload):
-        super().__init__(upload)
-
-        # multiext fits usually encode metadata, which we use to resolve which
-        # standardizer to use, in primary header. Primary header does not usually
-        # contain an image. Only image headers contain valid WCSs. Image headers
-        # do not, however, contain enough metadata to resolve which standardizer
-        # to use.
-        self.primary = self.hdulist["PRIMARY"].header
-        self.standardizer = HeaderStandardizer.fromHeader(self.primary,
-                                                          filename=upload.filename)
-
-        self.exts = []
-        for hdu in self.hdulist:
-            if self._isImageLikeHDU(hdu):
-                self.exts.append(hdu)
-
-    @staticmethod
-    def _isImageLikeHDU(hdu):
-        if not any((isinstance(hdu, CompImageHDU), isinstance(hdu, PrimaryHDU),
-                    isinstance(hdu, ImageHDU))):
-            return False
-
-        # People store all kind of stuff even in ImageHDUs, let's make sure we
-        # don't crash the server by saving 120k x 8000k table disguised as an
-        # image (I'm looking at you SDSS!)
-        if hdu.data is None:
-            return False
-
-        if len(hdu.data.shape) != 2:
-            return False
-
-        if hdu.shape[0] > 6000 or hdu.shape[1] > 6000:
-            return False
-
-        return True
-
-    @classmethod
-    def canProcess(cls, upload, returnHdulist=False):
-        canProcess, hdulist = super().canProcess(upload, returnHdulist=True)
-        canProcess = canProcess and cls._isMultiExtFits(hdulist)
-        if returnHdulist:
-            return canProcess, hdulist
-        return canProcess
-
-    def standardizeWcs(self):
-        standardizedWcs = []
-        for i, ext in enumerate(self.exts):
-            standardizedWcs.append(self.standardizer.standardizeWcs(hdu=ext))
-        return standardizedWcs
-
-    def standardizeHeaderMetadata(self):
-        return self.standardizer.standardizeMetadata()
-
-    def standardizeHeader(self, short=True):
-        standardizedWcs = {"wcs" : self.standardizeWcs()}
-        standardizedMetadata = {"metadata" : self.standardizeHeaderMetadata()}
-
-        if short:
-            standardizedMetadata.update(standardizedWcs)
-            return standardizedMetadata
-        else:
-            standardizedHeader = {}
-            for wcs in standardizedWcs["wcs"].values():
-                standardizedHeader[key] = dict(standardizedMetadata["metadata"],
-                                               **wcs)
-            return standardizedHeader
-
-    def storeThumbnails(self):
-        for i, ext in enumerate(self.exts):
-            large, small = self._createThumbnails(self._upload.basename + f"_ext{i}",
-                                                  ext.data)
-        plt.imsave(small["savepath"], small["thumb"])
-        plt.imsave(large["savepath"], large["thumb"])
-
-    @transaction.atomic
-    def storeHeaders(self):
-        # TODO: need to standardize the schema to be able to implement this
-        pass
