+++ trail/upload/process_uploads/processors/__init__.py
@@ -1,5 +1,6 @@
 # This import is required to trigger init_subclass in UploadProcessor
-from .fits_processors import *
+from .single_extension_fits import *
+from .multi_extension_fits import *
 from .rubin_processor import *
 from .decam_processor import *
